Update Jhipster 7.2.0
-> update version jhipster 7.2.0
-> update version lib javascript (...)
-> update version lib java (...)
-> update version image docker (...)
-> update BaseSearch
-> regenerate code from catiny.jdl
-> update Jenkinsfile (task parallel)

+++ Jenkinsfile
@@ -1,52 +1,47 @@
 #!/usr/bin/env groovy
 
-node {
-  stage('checkout') {checkout scm}
-
+node
+{
+  currentBranch = scm.branches[0].name
   currentFolder = sh(script: 'pwd', , returnStdout: true).trim()
   timeSleepingBeforePrintLog = 120
   timeSleepingWaitMariadbReady = 120
 
+  stage('checkout') {checkout scm}
+
   stage('check')
   {
-    def stageParallel = [:]
-    stageParallel["java"] =
-    {
-      try {sh "java -version"}
-      catch (e)
+    parallel(
+      'java' :
       {
-        echo "not installed java"
-      }
-    }
-    stageParallel["node"] =
-    {
-      try {sh "node -v"}
-      catch (e)
+        try {sh "java -version"}
+        catch (e) { echo "not installed java" }
+      },
+      'docker-compose' :
       {
-        echo "not installed node"
-        sh "curl -fsSL https://deb.nodesource.com/setup_lts.x | bash -"
-        sh "apt-get install -y nodejs"
-        sh "node -v"
-      }
-    }
-    stageParallel["docker-compose"] =
-    {
-      try {sh "docker-compose -v"}
-      catch (e)
+        try {sh "docker-compose -v"}
+        catch (e)
+        {
+          sh 'curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose && '+
+             'chmod +x /usr/local/bin/docker-compose && '+
+             'ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose'
+        }
+      },
+      'node' :
       {
-        echo "not installed docker-compose"
-        sh 'curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose && '+
-           'chmod +x /usr/local/bin/docker-compose && '+
-           'ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose'
+        try {sh "node -v"}
+        catch (e) { sh "curl -fsSL https://deb.nodesource.com/setup_lts.x | bash - && apt-get install -y nodejs && node -v" }
       }
-    }
-    parallel(stageParallel)
+    )
   }
 
-  stage('clean and nohttp') {
+  stage('clean and nohttp')
+  {
     sh "chmod +x gradlew"
-    sh "./gradlew clean --no-daemon"
-    sh "./gradlew checkstyleNohttp --no-daemon"
+    parallel(
+      "clean" : { sh "./gradlew clean --no-daemon" },
+      "no http" : { sh "./gradlew checkstyleNohttp --no-daemon" }
+    )
   }
 
 //   stage('npm install')
@@ -55,15 +50,16 @@ node {
 //   }
 //
 // Todo
-//    stage('backend tests') {
-//      try {
+   stage('backend tests') {
+     try {
+       sh "./gradlew test -PnodeInstall -Pdev --no-daemon"
 //        sh "./gradlew test integrationTest -PnodeInstall --no-daemon"
-//      } catch (err) {
-//        throw err
-//      } finally {
-//        junit '**/build/**/TEST-*.xml'
-//      }
-//    }
+     } catch (err) {
+       throw err
+     } finally {
+       junit '**/build/**/TEST-*.xml'
+     }
+   }
 // Todo
 //    stage('frontend tests') {
 //      try {
@@ -75,13 +71,9 @@ node {
 //      }
 //    }
 
-  stage('build and publish')
+  stage('build')
   {
-    sh "./gradlew bootJar jibDockerBuild -Pprod -PnodeInstall --no-daemon" // only build
-    docker.withRegistry('','docker-hub')
-    {
-      docker.image('yuvytung/catiny:latest').push()
-    }
+    sh "./gradlew bootJar jibDockerBuild -Pprod --no-daemon" // only build
   }
 
   stage('deploy')
@@ -107,45 +99,45 @@ node {
   			echo 'Sleep for ${timeSleepingWaitMariadbReady} seconds to wait for the mariadb to be ready'
   			sleep(timeSleepingWaitMariadbReady)
   		}
-
       sh "docker-compose -f ${currentFolder}/src/main/docker/app-dev.yml down catiny-dev-app"
       sh "docker-compose -f ${currentFolder}/src/main/docker/app-dev.yml up -d catiny-dev-app"
       sleep(timeSleepingBeforePrintLog)
       sh "docker logs docker_catiny-dev-app_1"
     }
-    stageParallel["prod"] =
-    {
-      try
-    	{
-    		sh 'docker inspect docker_catiny-mariadb_1'
-  //   		sh 'docker inspect docker_catiny-ffmpeg_1'
-    		sh 'docker inspect docker_catiny-redis_1'
-    		sh 'docker inspect docker_catiny-elasticsearch_1'
-  //   		sh 'docker inspect docker_catiny-kibana_1'
-    		sh 'docker inspect docker_kafka_1'
-    		sh 'docker inspect docker_zookeeper_1'
-    		sh 'docker inspect docker_jhipster-registry_1'
-    	}
-  		catch (ignored)
-  		{
-  			echo 'the necessary services (prod) are not running . try start it'
-  			sh "docker-compose -f ${currentFolder}/src/main/docker/app-prod.yml up -d"
-  			echo 'Sleep for ${timeSleepingWaitMariadbReady} seconds to wait for the mariadb to be ready'
-  			sleep(timeSleepingWaitMariadbReady)
-  		}
-      sh "docker-compose -f ${currentFolder}/src/main/docker/app-prod.yml down catiny-app"
-      sh "docker-compose -f ${currentFolder}/src/main/docker/app-prod.yml up -d catiny-app"
-      sleep(timeSleepingBeforePrintLog)
-      sh "docker logs docker_catiny-app_1"
-    }
+    if (currentBranch.equals("master"))
+      stageParallel["prod"] =
+      {
+        try
+      	{
+      		sh 'docker inspect docker_catiny-mariadb_1'
+    //   		sh 'docker inspect docker_catiny-ffmpeg_1'
+      		sh 'docker inspect docker_catiny-redis_1'
+      		sh 'docker inspect docker_catiny-elasticsearch_1'
+    //   		sh 'docker inspect docker_catiny-kibana_1'
+      		sh 'docker inspect docker_kafka_1'
+      		sh 'docker inspect docker_zookeeper_1'
+      		sh 'docker inspect docker_jhipster-registry_1'
+      	}
+    		catch (ignored)
+    		{
+    			echo 'the necessary services (prod) are not running . try start it'
+    			sh "docker-compose -f ${currentFolder}/src/main/docker/app-prod.yml up -d"
+    			echo 'Sleep for ${timeSleepingWaitMariadbReady} seconds to wait for the mariadb to be ready'
+    			sleep(timeSleepingWaitMariadbReady)
+    		}
+        sh "docker-compose -f ${currentFolder}/src/main/docker/app-prod.yml down catiny-app"
+        sh "docker-compose -f ${currentFolder}/src/main/docker/app-prod.yml up -d catiny-app"
+        sleep(timeSleepingBeforePrintLog)
+        sh "docker logs docker_catiny-app_1"
+      }
     parallel(stageParallel)
   }
 
-  stage('quality analysis')
+  stage('publish and quality analysis ')
   {
-    withSonarQubeEnv('sonar')
-    {
-      sh "./gradlew sonarqube -Pprod -PnodeInstall --no-daemon"
-    }
+    parallel(
+      'publish' : { docker.withRegistry('','docker-hub') { docker.image('yuvytung/catiny:latest').push() } },
+      'quality analysis' : { withSonarQubeEnv('sonar'){sh "./gradlew sonarqube -Pprod -PnodeInstall --no-daemon"} }
+    )
   }
 }
